Revert "Stuff"
This reverts commit f2b80d191a344ec98301f9ebbe1740dadd6af338.

+++ src/main/java/cn/nukkit/network/protocol/PurchaseReceiptPacket.java
@@ -17,8 +17,8 @@ public class PurchaseReceiptPacket extends DataPacket {
 
     @Override
     public void decode() {
-        for (int i = 0, count = (int) this.getUnsignedVarInt(); i < count; i++) {
-            this.receipts.add(this.getString());
+    	for (int i = 0, count = this.getUnsignedVarInt(); i < count; i++) {
+    	    this.receipts.add(this.getString());
         }
     }
